--- src/StartScreen.js
@@ -1,9 +1,0 @@
-export default function StartScreen({ numQuestions }) {
-  return (
-    <div className="start">
-      <h2>Welcome to The React Quiz!</h2>
-      <h3>{numQuestions} questions to test your React mastery</h3>
-      <button className="btn btn-ui">Let's start</button>
-    </div>
-  );
-}
