+++ src/components/Options.js
@@ -1,0 +1,11 @@
+export default function Options({ question }) {
+  return (
+    <div className="options">
+      {question.options.map((option) => (
+        <button className="btn btn-option" key={option}>
+          {option}
+        </button>
+      ))}
+    </div>
+  );
+}
