+++ src/components/Question.js
@@ -1,0 +1,12 @@
+import Options from "./Options";
+
+export default function Question({ question }) {
+  console.log(question);
+
+  return (
+    <div>
+      <h4>{question.question}</h4>
+      <Options question={question} />
+    </div>
+  );
+}
