+++ src/components/StartScreen.js
@@ -1,0 +1,14 @@
+export default function StartScreen({ numQuestions, dispatch }) {
+  return (
+    <div className="start">
+      <h2>Welcome to The React Quiz!</h2>
+      <h3>{numQuestions} questions to test your React mastery</h3>
+      <button
+        onClick={() => dispatch({ type: "start" })}
+        className="btn btn-ui"
+      >
+        Let's start
+      </button>
+    </div>
+  );
+}
